+++ src/main/java/pe/edu/upc/trabajogrupo2/servicesinterfaces/IRolesService.java
@@ -6,8 +6,8 @@ import java.util.List;
 public interface IRolesService {
     public List<Roles> List();
     public void insert(Roles roles);
-    public void delete(int id);
+    public void delete(Long id);
     public void update(Roles roles);
-    public Roles ListId(int id);
+    public Roles ListId(Long id);
 
 }
